Move SubgraphWalker completely to cpg-core
